+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## unreleased
+
+* Updated `eslint` to `^9.19.0` and adopted flat config.
+
 ## 7.0.0-alpha.4 (January 17, 2025)
 
 * Updated all dependencies to the latest including React 19.